chore: product-endpoint-integration-test
creates first integration test file using supertest
refactors product route file

+++ src/db/mock-db-connection.ts
@@ -1,15 +1,17 @@
-import mongoose from 'mongoose';
+/* eslint-disable no-console */
 import { MongoMemoryServer } from 'mongodb-memory-server';
+import mongoose from 'mongoose';
 
 let mongoServer: null | MongoMemoryServer = null;
 //connect to mock db
-async function connect() {
+async function connect(): Promise<void> {
   mongoServer = await MongoMemoryServer.create();
   const uri = mongoServer.getUri();
-  await mongoose.connect(uri, { maxPoolSize: 10 });
+  console.log({ uri });
+  await mongoose.connect(uri, { maxPoolSize: 1 });
 }
 //disconnect and close connection
-async function disconnect() {
+async function disconnect(): Promise<void> {
   await mongoose.connection.dropDatabase();
   await mongoose.connection.close();
   if (mongoServer) {
@@ -18,12 +20,14 @@ async function disconnect() {
 }
 
 //clear the db, remove all data
-async function clearDatabase() {
+async function clearDatabase(): Promise<void> {
   if (mongoServer) {
-    const collections = await mongoose.connection.db.collections();
-    for (let collection of collections) {
-      const { deletedCount } = await collection.deleteMany({});
-      console.log(`Deleted docs from collection ${deletedCount}`);
+    const collections = await mongoose.connection.db?.collections();
+    if (collections) {
+      for (const collection of collections) {
+        const { deletedCount } = await collection.deleteMany({});
+        console.log(`Deleted docs from collection ${deletedCount}`);
+      }
     }
   }
 }
